+++ src/checks.rs
@@ -10,7 +10,7 @@ mod unordered_keys;
 
 // This trait is used for checks which needs to know of only a single line
 trait Check {
-    fn run(&mut self, line: LineEntry) -> Option<Warning>;
+    fn run(&mut self, line: &LineEntry) -> Option<Warning>;
 }
 
 // Checklist for checks which needs to know of only a single line
@@ -36,8 +36,7 @@ pub fn run(lines: Vec<LineEntry>) -> Vec<Warning> {
         }
 
         for ch in &mut checks {
-            // TODO: Use a reference instead of the clone method
-            if let Some(warning) = ch.run(line.clone()) {
+            if let Some(warning) = ch.run(&line) {
                 warnings.push(warning);
             }
         }